pyproject.toml
[project]
name = "uv-fastapi-template"
version = "0.1.0"
description = "A minimal FastAPI template with uv dependency manager"
authors = [{ name = "Author", email = "author@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "uvicorn[standard]>=0.24.0",
    "email-validator>=2.3.0",
    "python-multipart>=0.0.20",
    "ruff>=0.14.5",
]

dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ====== 新增以下配置 ======
[tool.hatch.build.targets.wheel]
packages = ["app"]
# =========================

[tool.uv]
